requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-telegram-bot==21.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pyyaml==6.0.2
aiofiles==24.1.0
httpx==0.27.2
